showed date posted

+++ app.py
@@ -67,7 +67,7 @@ def addData():
                 'productCategory': productCategory,
                 'price': float(price),
                 'imageURL': imgURL,
-                'stockNumber': float(stockNumber),
+                'stockNumber': int(stockNumber),
                 'featured': isFeatured,
                 'datePosted': dateToday 
             }
